+++ contracts/pool/IPoolCore.sol
@@ -41,7 +41,7 @@ interface IPoolCore is IPoolInternal {
             address quote,
             address oracleAdapter,
             UD60x18 strike,
-            uint64 maturity,
+            uint256 maturity,
             bool isCallPool
         );
